公众号agent/requirements.txt
# Deep Scholar AI - 依赖包列表

# 核心依赖
asyncio-throttle==1.0.2
aiofiles==23.2.1
aiohttp==3.9.0
pyyaml==6.0.1
jinja2==3.1.2
click==8.1.7

# 机器学习和AI
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0
numpy==1.24.3
scikit-learn==1.3.0

# Web API
fastapi==0.110.0
uvicorn==0.24.0

# 文本嵌入
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# 向量数据库
chromadb==0.4.15
qdrant-client==1.6.4

# 知识图谱
networkx==3.1
neo4j==5.13.0

# PDF处理
PyMuPDF>=1.26.3
pdfminer.six==20221105
nougat-ocr==0.1.17
pillow==10.0.1
pymupdf4llm==0.0.27

# 网络爬虫和数据获取
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
feedparser==6.0.10
tweepy==4.14.0
arxiv==1.4.8
watchdog==4.0.0

# 日志和监控
python-json-logger==2.0.7
structlog==23.2.0

# 数据库
sqlalchemy==2.0.23
sqlite-utils==3.35.1

# 工具库
python-dateutil==2.8.2
python-dotenv==1.0.0
typing-extensions==4.8.0

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0
mypy==1.7.0

# 可选依赖（用于特定功能）
# 如果不需要某些功能，可以注释掉对应的包

# CUDA支持（如果需要GPU加速）
# torch==2.1.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html

# 额外的PDF处理工具
# unstructured==0.10.28
# pdf2image==1.16.3
# pytesseract==0.3.10

# 高级网络爬虫
# selenium==4.15.2
# playwright==1.40.0

# 数据可视化
# matplotlib==3.7.2
# seaborn==0.12.2
# plotly==5.17.0

# 高性能计算
# numba==0.58.1
# cupy==12.2.0

# 监控和部署
# prometheus-client==0.19.0
# uvicorn==0.24.0
# fastapi==0.104.1
